Improved the gravity implementation. Added a Player and a Circle Enemy. Implemented an abstract shape class and two subclasses. Added collision detection for X movement.

+++ src/me/jack/LD35/Entity/Entity.java
@@ -20,7 +20,7 @@ public abstract class Entity{
     public void addY(float add){
         this.y +=add;
     }
-    public void addX(int x) {
+    public void addX(float x) {
         this.x += x;
     }
     public float getX() {
